requirements.txt
pytest
pytest-html
pytest-playwright
requests
